Add ServiceArea module providing servable method
Add new ServiceArea module with a method `servable?` in which we apply our specific criteria for deciding whether the postcode is servable. This makes use of PostcodesIoClient (which is generic and knows nothing about our particular use of that API). We look at the 'lsoa' field, and use that to provide the answer back to PostcodeController.

Add a config file config/servable_lsoa_prefixes.txt which essentially defines our service area. For now we will check this in. In future a checked-in example file with a different file name may be preferable.

TODO: Implement allowed postcode list to override this in special cases.