Add PostcodeAllowedList for override postcodes
Add a PostcodeAllowedList module which manages the list of postcodes which are allowed as an override, either because
they would not normally meet the service area (lsoa based) criteria, or because they are not known to postcodes.io

Read the list from a config/postcode_allowed_list.txt file. The configured postcodes are matched against input ignoring whitespace.