Improved postcodes.io error handling
Handle 500 errors as well as timeouts etc within the HTTP calling logic to trap these errors as a custom exception type PostcodesIoClient::Error and report an error to the user.

Handle the 404 'not found' postcode scenario with the same display as a non-servable postcode, but with a little extra message.